feat(app): fixing the types for reducers
