+++ src/app/auth/services/auth.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core';
 import { AuthBlService } from './auth-bl.service';
-import { UserModel } from '../models/auth.model';
 import { Store } from '@ngrx/store';
 import { AppState } from 'src/app/shared/shared.state';
 import { autoLogoutAction } from '../states/auth.actions';
+import { UserModel } from 'src/app/shared/shared.model';
 
 @Injectable({
   providedIn: 'root'
