+++ src/app/auth/states/auth.selectors.ts
@@ -1,6 +1,6 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store';
-import { AuthState } from './auth.state';
 import { StateEnum } from 'src/app/shared/shared.enum';
+import { AuthState } from 'src/app/shared/shared.model';
 
 const AUTH_STATE_CONST = createFeatureSelector<AuthState>(StateEnum.AUTH_STATE);
