+++ src/app/shared/shared.model.ts
@@ -1,3 +1,11 @@
+export class UserModel {
+    constructor(private email: string, private token: string, private localId: string, private tokenExp: Date) {}   
+    
+    get tokenExpDate(): Date {
+        return this.tokenExp;
+    }
+}
+
 export interface Post {
     id?: string;
     title: string;
@@ -7,4 +15,9 @@ export interface Post {
 export interface SharedApiState {
     loadingStatus: boolean;
     errorMessage: string;
+}
+
+export interface AuthState {
+    user: UserModel | null;
+
 }
