+++ src/app/shared/shared.state.ts
@@ -1,7 +1,6 @@
-import { SharedApiState } from './shared.model';
+import { AuthState, SharedApiState } from './shared.model';
 import { StateEnum } from './shared.enum';
 import {sharedApiReducer } from './shared.reducer';
-import { AuthState } from '../auth/states/auth.state';
 import { authReducer } from '../auth/states/auth.reducers';
 import { ActionReducerMap } from '@ngrx/store';
 
@@ -9,10 +8,10 @@ import { ActionReducerMap } from '@ngrx/store';
 
 export interface AppState {
     [StateEnum.LOADER_STATE]: SharedApiState;
-    [StateEnum.AUTH_STATE]: AuthState
+    [StateEnum.AUTH_STATE]: AuthState;
 }
 
-export const appReducer: ActionReducerMap<any> = {
+export const appReducer = {
     loaderState: sharedApiReducer,
     authState: authReducer
 };
